+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `--output` flag in `veritail generate-queries` now defaults to `queries.csv` instead of being required.
+
 ## [0.3.0] - 2026-02-28
 
 ### Changed